pyproject.toml
[tool.poetry]
name = "faker"
version = "0.1.0"
description = ""
authors = ["David Kelly <david@futuresbright.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^6.0.1"
pytest = "^8.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
